requirements.txt
certifi==2025.7.14
charset-normalizer==3.4.2
ckanapi==4.8
docopt==0.6.2
et_xmlfile==2.0.0
idna==3.10
numpy==2.3.1
openpyxl==3.1.5
pandas==2.3.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-slugify==8.0.4
pytz==2025.2
requests==2.32.4
setuptools==80.9.0
simplejson==3.20.1
six==1.17.0
text-unidecode==1.3
tzdata==2025.2
urllib3==2.5.0
wheel==0.45.1
